+++ include/data.h
@@ -2,5 +2,5 @@
 #define DATA_H
 #include <vector>
 std::vector<float> getVertices();
-std::vector<uint8_t> getIndices();
+std::vector<unsigned int> getIndices();
 #endif // !DATA_H
